backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
supabase
httpx
redis
python-dotenv
python-multipart
celery
pytest
asyncio
